% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr.R
\name{bkmr}
\alias{bkmr}
\title{Model Selection Using BKMR}
\usage{
bkmr(df, family.value = "binomial", iterations = 10000, inclusion = 0.5)
}
\arguments{
\item{df}{The dataset with the predictors and outcome.}

\item{family.value}{Default fits a binomial model.}

\item{iterations}{Number of iterations to run BKMR model. Default is 10000.}

\item{inclusion}{The \% for inclusion for posterior inclusion probability. Default is 0.5}
}
\value{
Returns the PIP for the dataset.
}
\description{
Takes in the generated dataset and fits a model using
bayesian kernel machine regression.
}
\examples{
bkmr(df, family.value = "binomial")




}
