% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_simulation_function.R
\name{combined_simulation_function}
\alias{combined_simulation_function}
\title{Title}
\usage{
combined_simulation_function(df, beta, formula, n_rep, sample_size, model)
}
\arguments{
\item{negative}{}
}
\description{
Summary
}
