% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_regression.R
\name{penalized_regression}
\alias{penalized_regression}
\title{Model Selection Using Ridge, LASSO, Elastic Net}
\usage{
penalized_regression(df, family.value = "binomial", model.type = "lasso")
}
\arguments{
\item{df}{The dataset with the predictors and outcome.}

\item{family.value}{Specifies the family for running the lasso model.
The default is family.value = "binomial".}

\item{model.type}{Specifies which model we're trying to run.
Possible values: "lasso", "ridge", "elastic net"}
}
\value{
Returns the coefficient dataframe for the dataset.
}
\description{
Takes in the generated dataset and fits a penalized
regression models using L2, L1 penalties.
}
\examples{
penalized_regression(df, family.value = "binomial",
alpha.value = 1)




}
