% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_x.R
\name{sample_x}
\alias{sample_x}
\title{Sample predictor values from dataset}
\usage{
sample_x(df, n)
}
\arguments{
\item{df}{The dataset you want to sample from.}

\item{n}{The number of rows you want in your sample.}
}
\value{
A dataframe of rows from the original dataset.
}
\description{
Sample n rows with replacement from the dataset
}
\examples{
data_x <- sample_x(data, 1000)



}
