% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Create simulation data}
\usage{
simulate_dataset(
  parameter_list,
  demographic_list,
  drug_list,
  mean_vec,
  covariance_matrix,
  threshold = 0
)
}
\arguments{
\item{parameter_list}{The vector of parameters specifying beta values.}

\item{demographic_list}{The list of parameters that specify the ratio of
different demographic values in your simulated dataset.}

\item{drug_list}{The names of the drugs that you want to include
in your dataset.}

\item{mean_vec}{Specify your mean vector for the simulated dataset.}

\item{covariance_matrix}{Specify your covariance matrix for the simulated dataset.}

\item{threshold}{The value of the threshold which below the values for the
drugs are set to 0. Default value is 0.}
}
\value{
A simulated dataset with the prespecified betas.
}
\description{
Creates new simulation data depending on the user specification for
mean and covariance.
}
