% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_x.R
\name{simulate_x}
\alias{simulate_x}
\title{Simulate predictor values for dataset}
\usage{
simulate_x(
  demographic_list,
  drug_list,
  mean_vec,
  covariance_matrix,
  threshold = 0,
  n
)
}
\arguments{
\item{demographic_list}{The list of parameters that specify the ratio of
different demographic values in your simulated dataset}

\item{drug_list}{The names of the drugs that you want to include
in your dataset}

\item{mean_vec}{Specify your mean vector}

\item{covariance_matrix}{Specify your covariance matrix}

\item{threshold}{The value of the threshold which below the values for the
drugs are set to 0. Default value is 0.}

\item{n}{The number of rows you want in your dataset}
}
\value{
A dataframe of rows for the generated dataset.
}
\description{
Simulate demographic covariates, drug, and drug interaction
}
\examples{
demo <- c("TXGROUP.4", "TXGROUP.8", "TXGROUP.16",
          "SEX.F", "RACE.Black", "RACE.Hispanic", "RACE.Other")
perct <- c(1/4, 1/4, 1/4, 1/2, 1/5, 1/8, 3/40)
pl <- data.frame(Demographic = demo, Percentage = perct)

drugs <- c("AAI", "BDG", "BIC", "YAC")
n_vars <- length(drugs)

mean_vector <- rep(0.5, n_vars)

rho <- 0.2
sigma_squared <- 0.24
cov_matrix <- matrix(nrow = n_vars, ncol = n_vars)
for (i in 1:n_vars) {
  for (j in 1:n_vars) {
    cov_matrix[i, j] <- rho^abs(i - j)
                      }
                    }
cov_matrix <- sigma_squared * cov_matrix

data_x <- simulate_x(demographic_list = pl, drug_list = drugs,
                     mean_vec = mean_vector, covariance_matrix = cov_matrix,
                     threshold = 0, n = 1000)
}
