% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{simulation_function}
\alias{simulation_function}
\title{Performs n simulations and model generation.}
\usage{
simulation_function(
  df,
  parameter_list,
  n_rep,
  beta_num,
  replace = TRUE,
  iterations = 10000,
  inclusion = 0.5
)
}
\arguments{
\item{df}{The original dataset.}

\item{parameter_list}{The vector of parameters specifying beta values and n samples.}

\item{n_rep}{The number of simulations to run per scenario.}

\item{beta_num}{The number of total possible betas in the dataset.}

\item{replace}{Whether you want to sample with or without replacement.
Default is replace = TRUE.}

\item{iterations}{Number of iterations to run BKMR model. Default is 10000.}

\item{inclusion}{The \% for inclusion for posterior inclusion probability. Default is 0.5}
}
\value{
a list with the following elements:
\itemize{
\item{coefficients}
\item{validation}
}
}
\description{
Takes in the original dataset and list of parameters and
perform n simulations as specified in the list of parameters.
}
\examples{
simulation_function(df = simulation, parameter_list = grid[1,])



}
