% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function_v1.R
\name{simulation_function_v1}
\alias{simulation_function_v1}
\title{Performs n simulations and model generation.}
\usage{
simulation_function_v1(
  parameter_list,
  n_rep,
  demographic_list,
  drug_list,
  mean_vec,
  covariance_matrix,
  threshold = 0,
  max_lambda = 10,
  file_paths
)
}
\arguments{
\item{parameter_list}{The vector of parameters specifying beta values and n samples.}

\item{n_rep}{The number of simulations to run per scenario.}

\item{demographic_list}{The list of parameters that specify the ratio of
different demographic values in your simulated dataset.}

\item{drug_list}{The names of the drugs that you want to include
in your dataset.}

\item{mean_vec}{Specify your mean vector for the simulated dataset.}

\item{covariance_matrix}{Specify your covariance matrix for the simulated dataset.}

\item{threshold}{The value of the threshold below which the values for the
drugs are set to 0. Default value is 0.}

\item{max_lambda}{Specifies the maximum lambda considered for LASSO.
Default is 10.}

\item{file_paths}{The output file names to add results to. Needs to be two files:
one for the coefficients from the original model, one with the validation
information.}
}
\description{
Creates a new simulated dataset each time and performs model selection.
}
