% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_function.R
\name{validation_function}
\alias{validation_function}
\title{Compute the MSE & Sensitivity & Specificity of the simulation.}
\usage{
validation_function(parameter_list, simulation_list)
}
\arguments{
\item{parameter_list}{The vector of parameters specifying beta values.}

\item{simulation_list}{The vector of parameters from the simulation}
}
\description{
Takes in the prespecified parameter vector and compares it
to the simulated one using mean squared error.
}
