% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_function.R
\name{simulation_function}
\alias{simulation_function}
\title{Performs n simulations and model generation.}
\usage{
simulation_function(df, parameter_list, n_rep, beta_num)
}
\arguments{
\item{df}{The original dataset.}

\item{parameter_list}{The vector of parameters specifying beta values and n samples.}

\item{n_rep}{The number of simulations to run per scenario.}

\item{beta_num}{The number of total possible betas in the dataset.}
}
\value{
a list with the following elements:
\itemize{
\item{coefficients}
\item{validation}
}
}
\description{
Takes in the original dataset and list of parameters and
perform n simulations as specified in the list of parameters.
}
\examples{
simulation_function(df = simulation, parameter_list = grid[1,])



}
